Extract status progress bar
